Extract a base class for setting up additional Develocity plugin fixtures
